airflow/requirements.txt
feedparser==6.0.10
newspaper3k==0.2.8